few changes

+++ src/users/users.controller.spec.ts
@@ -1,20 +1,23 @@
 import { PrismaService } from '@/prisma.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersController } from './users.controller';
+import { UsersService } from './users.service';
 
 describe('UsersController', () => {
   let prisma: PrismaService;
   let controller: UsersController;
   let userId;
+  let cookies;
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [UsersController, PrismaService],
+      providers: [PrismaService, UsersService],
+      controllers: [UsersController],
     }).compile();
-    prisma = module.get<PrismaService>(PrismaService);
     controller = module.get<UsersController>(UsersController);
-  });
-  beforeAll(async () => {
-    await prisma.user.deleteMany();
+    prisma = module.get<PrismaService>(PrismaService);
+    if (userId === undefined) {
+      await prisma.user.deleteMany();
+    }
   });
   it('Should create a user', async () => {
     const user = await controller.signup({
@@ -26,6 +29,16 @@ describe('UsersController', () => {
       age: 1,
       username: 'test',
     });
+    userId = user.id;
     expect(user.email).toBe('test@test.com');
+  });
+
+  it('Should login a user', async () => {
+    const user = await controller.login({
+      username: 'test',
+      password: 'test',
+    });
+    console.log(user);
+    expect(user.msg).toBe('User logged in');
   });
 });
